Refactor bad cell correction
Use function pointer to choose which correction scheme to use.
In this way, we can avoid if statement in the loop that prevents
the function call to be inlined. In addition, current version resolves
a previous issue that a std::string argument prevents vectorization
due to a possible throwing of exception.